Verify emitted code for union vs. bitshifting
Problem:
I need to be able to treat a u16 register as two u8 registers vice
versa. When using bitshifting, the code is portable but might not be
optimal, whereas with unions, I'd expect the code to be optimal but
perhaps not portable.

Solution:
Write a small test program to check which is better. Turns out, both
produce similar optimal code (with -O2) and hence I can go with the
portable bit-shifting option.